Add section of code that plays a round
This section is an infinite loop that will break when all the board
positions are filled or when there is a winner. The player choices are
stored in an array variable of the instance itself, and at the same time
saved in a collection of the game instance for comparison purposes. The
set of selections per player is searched in an array of the Game class
with the results, with the help of the include? method, if this returns
true the game is over