Fix the code logic that decided the round winner
The code only returned true when the player's chosen positions were
equal to a subset of the Game class's results array. Therefore when the
array of chosen positions was greater than 3, despite containing the
winning positions, they were not equal to any sub-array of the results.

The solution was to add a second evaluation where the check was done by
intersection evaluation, the returned value is compared with the
sub-matrix of the results.