Add ComputerPlayer class
This class is responsible for creating the machine player instances.
This object has method that creates the array code randomly to be
guessed by  the human player. The code must contain 4 colors, which are
randomly  selected from the COLORS array from Player class. A criterion
that must be met by the code is that a color can only be repeated at
most twice within the array