Add check answer method
This method will get the answer given by the human user and compare that
solution with its own code, store the feedback and at the end give it to
the user. feedback isa hint based on the given answer.

The red pin will be returned when a colored piece is part of the code
and is positioned exactly the same as in the code. The white pin will
be returned when the colored piece is part of the code, but is in a
different position or index