Fix check_method code logic
Previously in the code, the code array was assigned to a variable in
order to store the value and manipulate it. The problem is that this
variable was a reference to the instance variable. So the code was
destroyed in a single round. For this, the enumerable map method was
used to iterate over the instance variable code and return an identical
array and store it in a new variable that can be manipulated