Configure logic within the Game class file
This Game class file requires the ComputerPlayer and HumanPlayer class
files. The creation of instances of the previous two classes will be
stored within instance variables in the Game class. The variables are to
reference the encoder object and the code guesser. Additionally, the
method that starts and controls the flow of the game was added.