Improve terminal line outputs
Changed some output strings to improve user experience. In addition, two
methods were added, the game_over method? which finishes executing the
loop when the player guesses the code. The other method is print_result,
as the name says it will print the game result based on the @winner
instance variable