Refactor code to allow human player to be coder
The code is refactored to allow the human user to decide whether they
want to be the code creator or the guesser. For this, a set_game method
is added where the user will be asked to decide the game mode option.
This option will be passed to another method that will be responsible
for creating the HumanPlayer and ComputerPlayer objects, which will be
stored or referenced in the instance variables that correspond to the
player roles in the game.