Add set_code and check_method methods
Now that the user can choose to be the creator of the code, the
necessary instance methods must be created for the Game class code flow
to work as it should. Added two methods, the first set_code simply asks
the user to enter 4 colors respecting the rules. The other method is
check_code, which receives the guess generated by the computer and the
user must enter the corresponding pins.