Add review method and refactor set_guess method
A new instance variable @final_guess is added, which will be an array of
size 4, which will be responsible for updating those positions that
match the secret code, so when a specific piece is guessed, it will be
passed to @final_guess in their respective index.

The review(obs) method will take as an argument the instance variable
|remark| of the HumanPlayer object and iterate over it, only values that
are different than nil will be updated in the variable @final_guess.

Additionally, the set_guess code was refactored, which previously, no
matter what, created a matrix of size 4 with random values as long as a
color was not repeated twice. Now @guess will initially take the value
of @final_guess, and will still add random colors only those positions
whose value is nil