Add ComputerPlayer method review(obs)
This method is responsible for passing the @remark variable of the
HumanPlayer object. This variable stores an array of size four, which
will be updated with those values ​​that are guessed by the computer.

The ComputerPlayer method will iterate over the given array and update
its variable @final_guess, this when the value is not nil