Add logic to give user to decide if play new game
Two methods were added to make a new game start at the end of the game,
if the user decides so.  The play_again method?  it simply obtains and
returns the user's decision. The game_flow method is an infinite loop
that creates the Game object and invokes the start method, then at the
end it invokes the play_again? method, if the value returned by the user
 is "n" a false is obtained and the loop ends