Add random word selection method
This method is executed when a new object of the class is instantiated.
The method reads the dictionary text file and stores each line in an
array object. Then a word is obtained randomly with the sample method.
This word will be returned only if it is greater than 5 or less than 12
characters.