Add method of reviewing given letter
This method iterates over the string secret_word and evaluates if the
given letter matches the character of the iteration. If the equality is
true, it will add the letter to the player variable guessed_letters and
the letter at that index will also be changed in the var secret_code. If
the letter does not belong to the string, it will be added to the player
variable wrong_guesses