Add logic to save and load game
Two methods were created for the game flow, one contains the logic to
save the game progress and finish the game. The method #save_game is
executed in each try of guessing, it waits for a confirmation from the
player. If a true value is obtained, it will serialize the instance of
the Game Class, as well as the player object that is part of it. This
will be done through a "#to_yaml" method, which will return the string
that will immediately be printed in a file with a .yaml extension. At
the end the game will stop and part of the state of the objects will be
printed.

The second method #load_game will be executed when starting the game and
if the file with a .yaml extension has a size. The player will be asked
if he wants to load the game, if 'y' is returned, the file will be read
and the string obtained will be deserialized with #from_yaml method,
passing the hash values to be able to build our Game instance and the
player object that is part of it. Then the flow of the game continues
to be the same.